SP Mod exp cast fix

+++ wolfcrypt/src/sp_armthumb.c
@@ -3168,7 +3168,7 @@ static int sp_2048_mod_exp_32(sp_digit* r, sp_digit* a, sp_digit* e,
         c -= bits % 4;
         if (c == 32)
             c = 28;
-        y = n >> c;
+        y = (int)(n >> c);
         n <<= 32 - c;
         XMEMCPY(r, t[y], sizeof(sp_digit) * 32);
         for (; i>=0 || c>=4; ) {
@@ -3312,7 +3312,7 @@ static int sp_2048_mod_exp_32(sp_digit* r, sp_digit* a, sp_digit* e,
         c -= bits % 5;
         if (c == 32)
             c = 27;
-        y = n >> c;
+        y = (int)(n >> c);
         n <<= 32 - c;
         XMEMCPY(r, t[y], sizeof(sp_digit) * 32);
         for (; i>=0 || c>=5; ) {
@@ -3991,7 +3991,7 @@ static int sp_2048_mod_exp_64(sp_digit* r, sp_digit* a, sp_digit* e,
         c -= bits % 4;
         if (c == 32)
             c = 28;
-        y = n >> c;
+        y = (int)(n >> c);
         n <<= 32 - c;
         XMEMCPY(r, t[y], sizeof(sp_digit) * 64);
         for (; i>=0 || c>=4; ) {
@@ -4135,7 +4135,7 @@ static int sp_2048_mod_exp_64(sp_digit* r, sp_digit* a, sp_digit* e,
         c -= bits % 5;
         if (c == 32)
             c = 27;
-        y = n >> c;
+        y = (int)(n >> c);
         n <<= 32 - c;
         XMEMCPY(r, t[y], sizeof(sp_digit) * 64);
         for (; i>=0 || c>=5; ) {
@@ -4997,7 +4997,7 @@ static int sp_2048_mod_exp_2_64(sp_digit* r, sp_digit* e, int bits,
         c -= bits % 5;
         if (c == 32)
             c = 27;
-        y = n >> c;
+        y = (int)(n >> c);
         n <<= 32 - c;
         sp_2048_lshift_64(r, norm, y);
         for (; i>=0 || c>=5; ) {
@@ -8779,7 +8779,7 @@ static int sp_3072_mod_exp_48(sp_digit* r, sp_digit* a, sp_digit* e,
         c -= bits % 4;
         if (c == 32)
             c = 28;
-        y = n >> c;
+        y = (int)(n >> c);
         n <<= 32 - c;
         XMEMCPY(r, t[y], sizeof(sp_digit) * 48);
         for (; i>=0 || c>=4; ) {
@@ -8923,7 +8923,7 @@ static int sp_3072_mod_exp_48(sp_digit* r, sp_digit* a, sp_digit* e,
         c -= bits % 5;
         if (c == 32)
             c = 27;
-        y = n >> c;
+        y = (int)(n >> c);
         n <<= 32 - c;
         XMEMCPY(r, t[y], sizeof(sp_digit) * 48);
         for (; i>=0 || c>=5; ) {
@@ -9608,7 +9608,7 @@ static int sp_3072_mod_exp_96(sp_digit* r, sp_digit* a, sp_digit* e,
         c -= bits % 4;
         if (c == 32)
             c = 28;
-        y = n >> c;
+        y = (int)(n >> c);
         n <<= 32 - c;
         XMEMCPY(r, t[y], sizeof(sp_digit) * 96);
         for (; i>=0 || c>=4; ) {
@@ -9752,7 +9752,7 @@ static int sp_3072_mod_exp_96(sp_digit* r, sp_digit* a, sp_digit* e,
         c -= bits % 5;
         if (c == 32)
             c = 27;
-        y = n >> c;
+        y = (int)(n >> c);
         n <<= 32 - c;
         XMEMCPY(r, t[y], sizeof(sp_digit) * 96);
         for (; i>=0 || c>=5; ) {
@@ -10812,7 +10812,7 @@ static int sp_3072_mod_exp_2_96(sp_digit* r, sp_digit* e, int bits,
         c -= bits % 5;
         if (c == 32)
             c = 27;
-        y = n >> c;
+        y = (int)(n >> c);
         n <<= 32 - c;
         sp_3072_lshift_96(r, norm, y);
         for (; i>=0 || c>=5; ) {
